Error messages in the case of message sending problems
