+++ widget/tool/chatactions/chataction.h
@@ -42,33 +42,4 @@ protected:
     QString name, date;
 };
 
-class MessageAction : public ChatAction
-{
-public:
-    MessageAction(const QString &author, const QString &message, const QString &date, const bool &me);
-    virtual ~MessageAction(){;}
-    virtual QString getMessage();
-    virtual void setTextCursor(QTextCursor cursor) final;
-
-private:
-    QString message;
-};
-
-class FileTransferAction : public ChatAction
-{
-    Q_OBJECT
-public:
-    FileTransferAction(FileTransferInstance *widget, const QString &author, const QString &date, const bool &me);
-    virtual ~FileTransferAction();
-    virtual QString getMessage();
-    virtual void setTextCursor(QTextCursor cursor) final;
-
-private slots:
-    void updateHtml();
-
-private:
-    FileTransferInstance *w;
-    QTextCursor cur;
-};
-
 #endif // CHATACTION_H
